requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
blinker==1.8.2
cachetools==5.5.0
cbor==1.0.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
datasets==2.19.0
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.16.1
FlagEmbedding==1.3.3
Flask==3.0.3
frozenlist==1.5.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
inscriptis==2.5.0
ir_datasets==0.5.9
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.17
langchain-community==0.2.19
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
loguru==0.7.3
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.18.3
networkx==3.1
numpy==1.24.4
openai==1.57.4
openpyxl==3.1.5
orjson==3.10.12
packaging==24.2
pandas==2.0.3
peft==0.13.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
propcache==0.2.0
protobuf==5.29.1
psutil==6.1.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyproj==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.20.1
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
streamlit==1.36.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unlzw3==0.2.2
urllib3==2.2.3
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==4.0.2
Werkzeug==3.0.6
xlrd==2.0.1
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
zlib-state==0.1.9
